src/python/main/requirements.txt
PyYAML
selenium
requests>=2.0.0
webvtt-py>=0.4.6


